requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
pandas
sqlalchemy
rq
redis
playwright
python-dotenv
lxml
pyarrow
openpyxl
psycopg2-binary
python-magic
playwright-stealth
google-generativeai
matplotlib
seaborn
tabulate
plotly